requirements.txt
adversarial-robustness-toolbox==1.19.1
appnope==0.1.4
asttokens==3.0.0
Bottleneck @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_55txi4fy1u/croot/bottleneck_1731058642212/work
certifi==2025.8.3
charset-normalizer==3.4.2
comm==0.2.2
contourpy @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_2cvjf0v4ux/croot/contourpy_1732540055997/work
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
eagerpy==0.30.0
executing==2.2.0
filelock==3.18.0
fonttools==4.56.0
foolbox==3.3.4
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.45
idna==3.10
ipykernel==6.29.5
ipython==9.0.1
ipython_pygments_lexers==1.1.1
jedi @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_38ctoinnl0/croot/jedi_1733987402850/work
Jinja2==3.1.6
joblib @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/joblib_1728589613021/work
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_cc2l_z_0ri/croot/kiwisolver_1737039586949/work
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mlxtend==0.23.4
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.5
numba==0.61.2
numexpr @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/numexpr_1731706619640/work
numpy @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_c1bynexs0d/croot/numpy_and_numpy_base_1742570425060/work/dist/numpy-2.2.4-cp313-cp313-macosx_11_0_arm64.whl#sha256=1e181012139473df297ef2682717f58a85878caf05580c9c795f5605016a9d25
packaging @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_a6_qk3qyg7/croot/packaging_1734472142254/work
pandas @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_4aifrweohv/croot/pandas_1732735109535/work/dist/pandas-2.2.3-cp313-cp313-macosx_11_0_arm64.whl#sha256=8f01410ead9f51bf614097bf4230440a4461983f8acb57141e819b93d81e26d6
parso @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_8824a1w4md/croot/parso_1733963320105/work
pexpect==4.9.0
pillow @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_85xj2lenf1/croot/pillow_1738010251686/work
platformdirs==4.3.7
prompt_toolkit==3.0.50
psutil==7.0.0
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure_eval==0.2.3
Pygments==2.19.1
pynndescent==0.5.13
pyparsing==3.2.1
python-dateutil @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/python-dateutil_1728585579455/work
pytz @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/pytz_1728586101417/work
pyzmq==26.3.0
requests==2.32.4
scikit-learn @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_2coq3750ge/croot/scikit-learn_1737988773864/work
scipy @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_cc_7w74fto/croot/scipy_1743153255057/work/dist/scipy-1.15.2-cp313-cp313-macosx_11_0_arm64.whl#sha256=3a16199878e888a9185db694e6cd911c3be3e8a9855d92fabeea915e83731ec6
setuptools==72.1.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.2
stack-data==0.6.3
sympy==1.14.0
threadpoolctl @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_18rgexrkuo/croot/threadpoolctl_1731006101562/work
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tornado==6.4.2
tqdm @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_5642tzrprr/croot/tqdm_1738943467095/work
traitlets @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/traitlets_1728585482264/work
typing_extensions==4.14.1
tzdata @ file:///croot/python-tzdata_1690578112552/work
ucimlrepo==0.0.7
umap-learn==0.5.9.post2
urllib3==2.5.0
wcwidth==0.2.13
wheel==0.45.1
